Create CODEOWNERS to complete open source review.
Per comments from @mnagelia we ought to have a clear mechanism for code ownership, and the CODEOWNERS file has been recommended in other open-sourcing initiatives. Specifically, during final review, it was asked if possible to specify groups and not individual users assigned to contractor teams as this causes problems during staffing shuffles.